+++ dango/ics07-tendermint/src/ctx/mod.rs
@@ -1,6 +1,7 @@
 //! This module contains the context implementations required by ibc-rs.
 
+#[allow(clippy::module_name_repetitions)]
 pub mod client_ctx;
-mod ctx;
+mod tendermint_ctx;
 
-pub use ctx::*;
+pub use tendermint_ctx::*;
